Bugfix: avoid double encoder output to OSC
Avoid outputting two values via OSC from the encoders. One value
is output from zynthian-gui, and includes consideration to the
max value, stepping, etc. The other value is output from Zyncoder,
and is always 0..127 . This is not noticeable when the parameter
range is 0..127; both values output will be the same. But for instance
if a parameter is set up which has value range of 0,1,2,3 , then the
Zyncoder will output 0,32,64,96 and zynthian-gui will output 0,1,2,3,
confusing OSC recipient.

The solution is to simply not set up Zyncoder to output OSC values,
as zynthian-ui will do this anyway (and more correctly as noted above).

This was discovered while attempting to implement support for
ZynAddSubFX part parameter polyType (which replaces the deprecated
Ppolymode and Plegatomode). I do not know if this is a problem only
for ZynAddSubFX or any other OSC client (are there any others?).